ivc: fix missing peer notification
Ensure that we notify the remote end each time our local
state has changed. Previously, we neglected notificaton if
the remote end was already in the established state, but
this optimization can lead to a hang in the protocol if
the remote end is undergoing a state change.

VFND-1786

Change-Id: I1de31d8453ec7ba32d3b79bf4bd40f7edf41aeed
Signed-off-by: Dennis Kou <dkou@nvidia.com>
Reviewed-on: http://git-master/r/1149233
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vladislav Buzov <vbuzov@nvidia.com>
Reviewed-on: http://git-master/r/1198265
Tested-by: Haley Teng <hteng@nvidia.com>
Reviewed-by: Bahadir Balban <bbalban@nvidia.com>