dma-coherent: Remove alignment when client passes DMA_ATTR_ALLOC_EXACT_SIZE
If client passes DMA_ATTR_ALLOC_EXACT_SIZE, it is most likely
that it needs to mark the exact range it specified as occupied.
If alignment is non-zero in such a case, the region which is marked
as occupied is not what the client really intends to. So, keep
alignment requirement as 0 when DMA_ATTR_ALLOC_EXACT_SIZE is
specified.

TMM-40

Change-Id: Ic51d80a028f1ab6becbf21ce90620f7edfa8d9af
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/1157202
(cherry picked from commit e4a2992192209ecf70abb75b44071d9c47d08e2b)
Reviewed-on: http://git-master/r/1158173
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>