cpufreq/schedutil: Prepare util hook for remote updates
Pass in strct rq as an argument for the cpufreq_update_util so that the
rq->cpu's utilization hook can be fired. Currently the hook always updates
assuming it is called for the local cpu (this_cpu_*).

This is a preparatory patch to allow the hook to fire even for remote
updates.

TPM-955

Change-Id: I4aa818c53173622e5f1ac06880dca9fa782cd14a
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1164514
GVS: Gerrit_Virtual_Submit
Reviewed-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>