rtc: tegra: merge downstream and upstream driver
This merges in functionality from the 3.18 downstream
driver into the upstream driver.
An effort has been made to reduce the delta between the
two drivers while keeping all functionality from
the downstream driver.
One exception is an omission of the registration
of a persistant clock.

Bug 200193765

Change-Id: I95954fa3bfc0c6852a8283d08f6d6781d3971001
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: http://git-master/r/1164384
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>