spi: tegra: add support for delay between packets
Add support for delay between packets. SPI controller does
not support this feature directly but it supports the
chipselct inactive between packets and the number of clock
cycle for inactive state.

So to keep chipselect active between packets with delay, chipselect
is converted to GPIO mode and set for proper level and configure
controller to CS inactive between packets to get the packet delay.

Bug 1564116

Change-Id: I550a8109f4ae2340aaa37fc0f2d4141dd5e9e787
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Reviewed-on: http://git-master/r/1171588
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit