usb: gadget: xudc: toggle vbus if PLC to inactive
When doing quick cable unplug-plug with XUDC connected to a U2-enabled
host controller, it's normal that the port will enter inactive state if
the unplug-plug interval is so small that the extcon VBUS removal
notification doesn't trigger. When the VBUS removal notification doesn't
happen, port will stay in Inactive state.

This change start a 100ms timer when PLC to Inactive state. If the port
is still in Inactive state after 100ms (without CSC), we toggle VBUS so
the port won't be stuck in Inactive state.

Bug 200204478

Change-Id: I44b989144f2d18497a7f75b4b547f0593d38ac18
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1172898
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: ChihMin Cheng <ccheng@nvidia.com>
Reviewed-by: Rohith Seelaboyina <rseelaboyina@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>