Kbuild: relative path to tegra projects
Create a make function (tegra-path) that define the relative paths to
Tegra kernel projects. When the project is split out of the main kernel
directory, it resolves to someothing like:
../../<tegra-project>/drivers/tegra_stuff

But, when the project is integrated into the main kernel the same function
evaluates to:
../drivers/tegra_stuff

It does this by testing for ../../<tegra-project>.

Add the function tegra-base-kernel-path to be used in sound/

Automatically search for <somepath>/Makefile.<project> to handle:
    obj-y += <somepath>/

In this example, <project> is an NVidia specific external kernel project.

Bug 1783210

Change-Id: I020cb7bac5b33ef47b4d8cf07da39806fbd0bb07
Signed-off-by: dmitry pervushin <dpervushin@nvidia.com>
Reviewed-on: http://git-master/r/1174848
Reviewed-on: http://git-master/r/1177285
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
Tested-by: Alexander Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/1183039
Reviewed-by: Adeel Raza <araza@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>