input: gpio_keys: handle the missing key press event in resume phase
The GPIO key press event might be missed in the resume phase, if the key
had been released before the system had been resumed to the stage that it
could capture the press event. So we simulate the wakeup key press event
in case the key had been released by the time we got interrupt handler
to run.

Bug 200193765

Change-Id: Ic5e7f98ef7f880f2ac7ed78670113d417d044c28
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1175524
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Tested-by: Sivaram Nair <sivaramn@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>