input: gpio_keys: Set suspended flag in pm notify callback
Set suspended flag in pm notify callback for handling it at the
beginning of the suspend and at the end of the resume.
If the gpio key is routed through a gpio expander chip, the gpio key
irq handler could be entered later than the gpio keys resume callback.

Bug 200217026

Change-Id: I372302d008787d8619854e6b10dd3256014c2ddf
Signed-off-by: Jinyoung Park <jinyoungp@nvidia.com>
Reviewed-on: http://git-master/r/1183832
GVS: Gerrit_Virtual_Submit
Tested-by: Joseph Lo <josephl@nvidia.com>
Reviewed-by: Joseph Lo <josephl@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>