pinctrl: tegra: Set SFIO mode to Mux Register
A pin should be muxed to set to SFIO mode to be used
for pinmux operation.
So mux setting should set the Pinmux mode by default
also when the GPIO is freed/released pinmux mode is to
be set to ensure default behavior is restored.

Bug 200225318

Change-Id: I91386b288fb05a70918f7b87799976e412197557
Signed-off-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-on: http://git-master/r/1197417
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>