mmc: tegra: skip features not supported in VDK
In VDK based platforms, we don't support clocks yet. Hence, sdhci
probe will result in failure when it tries to get clocks.
Fix this by trying to not get clocks in VDK platforms.

We also don't support extended GP in VDK, even if EXT_CSD
indicates so. Mark the mmc capabilities flag to so that
we bail out from accessing extended GP.

Bug 1794876

Change-Id: I897d51bf5e4603edf82c047af9619d451647a5c3
Signed-off-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-on: http://git-master/r/1195437
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>