usb: gadget: tegra-xudc: fix charger detection
There are several problems in the original charger detection integration
in XUDC driver:

1. when system boots up with USB cable attached (to host), the charger
type will be erroneously detected as non-standard. Because XUDC probe()
is invoked earlier, BC protocol is carried out early too. When SDP is
determined, we are expecting the first setup packet to arrive in 1
second. If not, we set to non-standard charger type. In this case the
user space "init" program hasn't got a chance to configure gadget
functions to XUDC until a few seconds later, so we always end up with
non-standard charger type. Fix it by checking if xudc->pullup is already
set. If it's not, do not schedule the work to set charger type to
non-standard.

2. sometimes it may take more than 1 second for the setup packet to
arrive (gadget configuration is ready) so we erroneously set the charge
type to non-standard even if it should be SDP/CDP. Extend the timeout
value from 1 second to 2 seconds.

3. When device mode is turned off, we should also set charger type to
NONE by calling charger detection driver's API. This will ensure the
charger type is cleared when the USB cable is removed.

4. connect_type should be protected. Add spinlock to protect them from
concurrent access.

Bug 200148419
Bug 200227601

Change-Id: I2a51cdc6ebcaca442cb39a6450e24a9f8199e6d4
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1205690
Reviewed-by: Rakesh Babu Bodla <rbodla@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>