usb: gadget: usb_cd: implement VBUS wakelock
Implement VBUS wakelock that resembles the old otg-wakelock.c driver.
The feature is implemented here (in battery charger detection driver)
because this driver is the first place where the actual USB charger
type is determined. We utilize this information to determine when to
hold/drop a wakelock to prevent system from suspending.

The rules are:
1. hold wakelock whenever SDP/CDP type of charger cable is detected
   (which means when data connection is present)
2. hold a 2 second temporary wakelock whenever other type of power
   sources (ex. DCP/non-standard, etc) is detected
3. drop the wakelock whenever cable is disconnected

This feature is by default enabled and can be disabled by module
parameter "vbus_wakelock" of this tegra_usb_cd driver.

Note that the only client of this driver is the XUDC driver, which
registers an extcon cable notification to receive physical cable
connection/disconnection event. This functionality (wakelock) depends
on XUDC driver to know when to hold/drop the wakelock.

Bug 200148419
Bug 200227601

Change-Id: I34bca6282e719e6360660a55eae02fd33149e5d1
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1205691
Reviewed-by: Rakesh Babu Bodla <rbodla@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>