usb: gadget: tegra-xudc: fix squeeze ring logic
There is an error in the squeeze_transfer_ring() which may cause the
loop to be infinite.

When enqueue pointer is 0, and the initial value of i isn't 0, the loop
becomes infinite because the termination condition will never be met.
Since the code runs with XUDC spinlock held, it will cause CPU hard
lockup when the next USB dequeue request comes in.

Fix the error by using the while loop.

Bug 200223265

Change-Id: Icd2fca86ddd4757cadd16078358da435b50078a5
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1213588
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Henry Lin <henryl@nvidia.com>
Reviewed-by: Hans Yang <hansy@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>