script: Makefile: Do not create relative path for DTS temp files
When compiling the DTS files, do not create the directory
with relative path. This may create the temp files to be outside
of out folder.

Replace all double-dot ("..") from the path to "_ddot_" so that
directory for all temp file will be stay in the same folder.

To achieve this, add rule for dtc compilation.

bug 200217137

Change-Id: I831cb8957368f54e3acb51c14d17da51b70f4995
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1215415
Reviewed-by: Stefan Becker <stefanb@nvidia.com>
GVS: Gerrit_Virtual_Submit