net: eqos: set TWT timer value to 21usec
The MAC cannot start the transmission until the
wake-up time specified for the PHY expires.
This wake-up time is suppose to be auto-negotiated
and to be programmed in TWT timer. Since we don't
have any PHY register to read the value,we use 21usec
as default value for BRCM PHY as per the data sheet.

Bug 200233029

Change-Id: Ib3a23fa8303e59aed76c7a5085646e661b39cf24
Signed-off-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-on: http://git-master/r/1220069
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>