drivers: platform: fix denver_mca driver
If one Denver asserts for eg. in case of JSR:MTS MCA then the
other Denver core deadlocks. Therefore in case of an asserting
Denver core, we have to assume that the Denvers are gone and
hence we need to read the Denver MCA banks of the asserting core
from A57 using ARI.
The current implementation of the denver MCA driver doesn't handle this.

This patch :
- Adds support for handling JSR:MTS MCA (SERR1 Bank) by allowing A57
  cores to read the asserting Denver core's MCA banks through ARI
  interface.

- fix mca_trip so that it runs only on a Denver core

Bug 1796736

Change-Id: I4f2b5409d7d62033f4fe46158b13b929a66db1a6
Signed-off-by: Rohit Khanna <rokhanna@nvidia.com>
Reviewed-on: http://git-master/r/1220270
Reviewed-by: Richard Wiley <rwiley@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>