platform: tegra: deprecate APIs to change waketable runtime
Deprecate the APIs tegra_set_wake_source() and
tegra_disable_wake_source() which changes the wakeup table
in runtime.

Add APIs to update the gpio wakeup table with gpio base if it
is different than zero.

The table is populated statically and it can be updated by only
GPIO driver to provide the base.

bug 200237366

Change-Id: Id9e307737f690ea3d7e50a39996d75f98b1cfae3
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1225232
GVS: Gerrit_Virtual_Submit