clocksource: tegra: add Tegra210 timer support
Add support for the Tegra210 timer that runs at oscillator clock.
We need this timer to work as system timer (clock event device) to
replace the ARMv8 architected timer due to it can't survive across the
power cycle of the CPU core or CPUPORESET signal. So it can't be a wake
up source when CPU suspend in power down state.

Based on the work by: Antti P Miettinen <amiettinen@nvidia.com>

Bug 200237195

Change-Id: I09fba55c2a00940071466be8789c8e0b8fb2e576
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1231786
GVS: Gerrit_Virtual_Submit
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>