ARM: cpuidle: add support for Tegra210
Tegra210 supports some more advanced idle states that need to cope with
BPMP-L. So we enhance the generic ARM CPU idle driver by adding
"tegra210_cpu_pm_enter" and exit call to deal with that. If the state is
not tolerated, it will always fall back to C7 state.

Bug 200237195

Change-Id: I16af3be073bd9d92d13d45349a5f27c84cca753c
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1231791
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>