usb: gadget: usb_cd: deassert PD before doing CD
We should deassert PD before doing charger detection according to PG, or
charger detection result is unreliable. Similarly, we assert PD when
charger detection is done.

Those sequence is now in this pair of APIs:
- tegra_phy_xusb_utmi_pad_charger_detect_on()
- tegra_phy_xusb_utmi_pad_charger_detect_off()

Also fix the sequence in usb_cd poweroff so it is done in a reverse order.

Bug 200232958
Bug 200239773

Change-Id: I6fef6a9b519628056f7aa3f0f3cccb92b0b37804
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1232477
GVS: Gerrit_Virtual_Submit
Reviewed-by: Rakesh Babu Bodla <rbodla@nvidia.com>
Reviewed-by: BH Hsieh <bhsieh@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>