gpio: tegra: moving suspend function to syscore
When supporting deep sleep mode on Tegra which means except always-on
power partitions and SDRAM in self-refresh mode, other blocks and main
power will be shut off. The GPIO settings can't be retained across power
cycle. It will cause the resume flow be broken when some of the early
resume functions or IRQ accessing the device that depends on GPIO to work
normally. Moving the suspend/resume function from system to syscore to
resolve this.

Bug 1811733

Change-Id: Iafd1130053bf70bf9e3bed215f4a5bc5689445af
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1254287
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>