soc/tegra: pmc: add LP0/SC7 suspend support
The LP0 is a deep sleep mode for Tegra SoCs. When it goes into LP0 mode,
the PMC will shut off all power partitions that includes vdd_cpu &
vdd_core power rails except the PMC partition with RTC inside. The SDRAM
will be put into self-refresh mode. And the warm boot code is a must for
LP0 waking up.

The SC7 state is going to replace the legacy term of LP0 in Tegra210 and
the later coming chips. The functionality is same as LP0, but some of the
control will be moved from CPU to BPMP to complete the LP0 sequence. For
ex, the SDRAM self-refreshing code and CRAIL power gating control code.

The sequence of LP0 suspending:
* turning off L2 & L1 data caches and the MMU
* setting DPD, warm boot flag & address for warm boot code
* putting SDRAM into self-refresh mode
* trigger LP0 from PMC
* PMC trigger cpu_power_req & core_power_req to PMIC to shut off the
  vdd_cpu & vdd_core power rail
* turning off the power partitions in the SoC

The sequence of LP0 resuming:
* The boot ROM detects the warm boot flag then putting the SDRAM back
  to normal mode and jumping to warm boot code
* The warm boot code is executed from BPMP-L
* The warm boot code re-enables the PLLX and PLLP that are needs for
  resuming
* The warm boot code brings up CPU0
* Then releasing CPU0 to go and start with reset vector

Bug 1811733

Change-Id: I0992f876ed42eb4dab965e5e1562577fc4030515
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1254302
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>