of: thermal: tegra: add bindings of hw throttle
Add HW throttle configuration sub-node for soctherm, which is used to
describe the throttle event, and worked as a cooling device. The "hot"
type in thermal zone can be bound to this cooling devic, and trigger
the throttle function.

Signed-off-by: Wei Ni <wni@nvidia.com>
(cherry-picked from https://lkml.org/lkml/2016/5/11/249)

Bug 200233003

Change-Id: I4cca2f3dc75ea7b93157db6c4d21330b04f6e835
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: http://git-master/r/1243910
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>