usb: gadget: tegra-xudc: fix NULL dereference
When handling transfer completion in IRQ, we unlock the spinlock to give
back the USB request to gadget layer and then lock again. It is likely
that after the execution of tegra_xudc_req_done(), ep->desc is written
to NULL in __tegra_xudc_ep_disable() so we got NULL dereference in
usb_endpoint_xfer_control(ep->desc).

Fix it by checking if ep->desc is valid before accessing the structure.

Bug 1836950

Change-Id: I6277f1c1ee9c532f79410152ec89b13eaeea348c
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/1249525
GVS: Gerrit_Virtual_Submit
Reviewed-by: Hans Yang <hansy@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>