PM / Domains: fix redundant genpd_save_dev call in suspend_noirq
This is wrong behavior that invokes "genpd_save_dev" function in
pm_genpd_suspend_noirq, which will finally call the runtime_suspend
callback function of the device. And this is not the expected behavior of
regular suspend sequence in kernel and doesn't align with upstream
kernel.

Same as the resume_noirq flow. We should fix it. But some drivers can
not work with it, e.g. pcie, so using the same flag
GENPD_FLAG_PM_UPSTREAM to make it work with other downstream drivers.

Bug 1811733

Change-Id: I9c5f040b48423134bcd9adc9863ff29bcfa62082
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1254298
GVS: Gerrit_Virtual_Submit
Reviewed-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>