soc/tegra: Add thermal cooling support for CPU EDP
The CPU EDP driver is missing thermal cooling support because for
the Chromium v3.18 kernel this was implemented as a separate driver.
However, in order to re-use the existing device-tree bindings from
the kernel v3.10, add thermal cooling support to the CPU EDP as this
aligns with the v3.10 kernel.

Bug 1811732

Change-Id: I5b43aa448529c652fa5a90d06ad5e923fef1030c
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1255687
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Tested-by: Jon Mayo <jmayo@nvidia.com>