WAR: PM / Runtime: Add wait conditions in __pm_runtime_barrier
Sometimes Kernel waits completion of runtime PM operations indefinitely
in __pm_runtime_barrier() when restarting or shuting down.
It's a side effect by http://git-master/r/116092 which is for freezing
processes before restart or shutdown.
As a workaround to avoid this issue, added conditions to wait for completion
of runtime PM operations if system state is booting or running.

Bug 1827071

Change-Id: Id0e5c6779b8bb7e56e43c049564d92021408f183
Signed-off-by: Jinyoung Park <jinyoungp@nvidia.com>
Reviewed-on: http://git-master/r/1257668
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Kurt Yi <kyi@nvidia.com>