clk: tegra: move tegra_dvfs_set_dfll_range to cpufreq driver
The DVFS constraints apply to cclk_g, not to dfll clock itself. Therefor
set the CPU rail to DFLL mode in the cpufreq driver as that driver knows
when we source cclk_g from the DFLL. This also prevents the call from
happening before the DVFS layer is initialized.

Change-Id: Idcea614f4cf173f6cf11424301e8299554875052
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1261578
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit