soc/tegra: sysedp: Clean-up error messages
It is not mandatory that platforms have 'consumer data' for SYSEDP and
so to avoid printing unnecessary errors messages for platforms without
consumer data, make the message a debug message and not an error message
by default.

Furthermore, for platforms without 'consumer data' the registration of
any SYSEDP consumer will always fail and this is expected. Therefore
only print an error message when a consumer fails to register if there
is consumer data available. Finally add a error message to indicate if
the registering of the consumer fails.

Bug 1811732

Change-Id: Ia582cba85a9818b55ef3f2f4cf12f5f58895dcdb
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1265858
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>