kernel: arm64: Add "model name" for 64bit tasks
Currently we have restriction to display processor model name under
cpuinfo structure for 32 bit tasks only and thus, ubuntu unity center
application or the /proc/cpuinfo does not display processor model name
for 64 bit tasks.

With this we are removing the restriction of displaying model name and 
/proc/cpuinfo and ubuntu unity center can display the processor 
information to the user for 64 bit tasks.

Bug 200242461

Change-Id: Ia7c6dde58d799e7de1c45d608e4156d921f266b0
Signed-off-by: Ninad Malwade <nmalwade@nvidia.com>
Reviewed-on: http://git-master/r/1268144
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Bibek Basu <bbasu@nvidia.com>