soc: tegra: Build for Tegra
There is two configs for the Tegra SoC build
	ARCH_TEGRA and PLATFORM_TEGRA

soc/tegra get build for the ARCH_TEGRA but some of SoCs
have not enabled the ARCH_TEGRA and hence it can not be build.

Add the new config ARCH_OR_PLATFORM_TEGRA which is enabled for
all SoCs and build this folder for all SoCs.

The driver specific source will be build for ARCH_TEGRA to maintain
original build.

This helps to remove the duplication on driver per SoCs, unified the
kernel build for all SoCs.

bug 200260692

Change-Id: I2df6c67e7b30c77eba04bf58a0ff0ad9cf51dec7
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1269329
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shardar Mohammed <smohammed@nvidia.com>
Tested-by: Shardar Mohammed <smohammed@nvidia.com>
GVS: Gerrit_Virtual_Submit