soc/tegra: edp: Update GPU EDP clock
For Linux v4.4, the GPU clock used by the GPU DVFS device is the
'gpcclk' and not the 'gbus' clock (child of the 'gpcclk') as it was
in Linux v3.10. This difference is causing the GPU EDP device to fail
probing when setting up the f/v table because the GPU clock used by the
GPU EDP device does not match that used by the GPU DVFS device. Although
this could be fixed by updating the GPU DVFS clock to be 'gbus' as it
was in Linux v3.10, this may cause other issues in the NVGPU driver when
quering the max frequency supported for GPU DVFS. Therefore, update the
GPU EDP driver to use the 'gpcclk' instead of the 'gbus' clock. Please
note that because the 'gpcclk' is only registered as a clkdev clock, we
cannot currently look it up via device-tree and have to look it up using
the legacy clk_get_sys() API.

Bug 1811732

Change-Id: I050b492d5526b0acd46b3bc0ed27c67bbb92371f
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1273469
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>