clk: tegra: fix disable unused for clocks sharing enable bit
In case 2 clocks share an enable bit and one of them is enabled by a driver
and the other one is not, CCF will think it's enabled because it will only
look at the hw state. Therefor it will disable the clock and thus also disable
the other clock which was enabled. Work around this by not decrementing the
refcount if the disable is due to an unused clock rather than an explicit
clk_disable call. A better solution would probably be to maintain separate
enable bits in sw. This is a bit more invasive to implement though.

Change-Id: I77280ec6a2a4b943f05330f67e1859798c6de9a3
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1272673
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>
Tested-by: Shreshtha Sahu <ssahu@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>