tegra: powergate: add BPMP based power domain provider
This adds a generic_pm_domain provider by using the services
provided by BPMP.

As a final goal, it is meant to fully replace the tegra
powergate framework on chips with bpmp support.

In the meanwhile, to be able to live alongside the tegra powergate
framework, it only registers the domains that reference this driver
through the device tree, instead of registering all domains that BPMP
controls.

Bug 1853511

Change-Id: I25292885905d3d7b60848df635c15384aa612757
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: http://git-master/r/1273521
Tested-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>