clk: handle early debugfs files
If __clk_debugfs_add_file or clk_debugfs_add_file is called before the
CCF debugfs nodes are created, store the request in a list to be
processed when the CCF debugfs nodes are added. This allows for easier
addition of debugfs files by other parts of the kernel without having
to worry about init dependencies.

Change-Id: Ib243b70104fc9e6b8b836d6ad9b182487aca87db
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1274254
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>