clk: tegra: dfll: fix find_lut_index_for_rate for regulator
Fixing the bug that when we finished looking up the DVFS table, there
will be a voltage index to match with the request rate. But the voltage
may not just match with the regulator output, we need to fix it as when
we find the minimum voltage of the target index of the DVFS table.

Bug 200262988

Change-Id: I99c17cf2559e00b75d7598580c7a9077297223f7
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1278610
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>
Tested-by: Shreshtha Sahu <ssahu@nvidia.com>
Reviewed-by: Ninad Malwade <nmalwade@nvidia.com>
Tested-by: Ninad Malwade <nmalwade@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>