soc/tegra: pmc: Get rid of linux/platform/tegra/io-dpd.h
The io-dpd.h header declare the APIs legacy way of enabling IO
DPD on IO pads. The legacy APIs are moved to pmc.h as the
static inline to use the new framework and hence this header
is no more used.

Remove this header as this is unused.

bug 200260700

Change-Id: Ib86a342f7f666ad9e34a6f61914127a3b06a4b1a
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1279180
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>