cpufreq: tegra210: Fix setting of DFLL range
The function tegra_dvfs_set_dfll_range() should be called once we switch
to the DFLL for the CPU clock source. However, this is actually being
called only if we fail to switch in the CPU clock source to the DFLL
because it is called in the error. Fix this and clean-up the error path
for clk_set_parent() because it is a bit confusing how it is written.

Bug 1811732

Change-Id: Ia452fe9a34bf8846223b55b522d91b9835d6fbe3
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1281370
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>