misc: mods: refactor mods clock driver
Clock ids and phandles are now precompiled in a dts file which is
available on boot. This information no longer needs to be stored in
a table and attached to the mods node using a live update. The driver
is updated to directly get clock names and ids from this node.

Bug 1854183

Signed-off-by: Vivek Aseeja <vaseeja@nvidia.com>
Change-Id: I484c5df10673d00f0772e62bb7d375b2e4285077
Reviewed-on: http://git-master/r/1281530
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>