soc/tegra: pmc: Add register enums into IO pad soc data
The PMC driver is used for the multiple SOCs and each SoCs has the
different DPD register and bits for the IO pads. To support the
multiple SOCs, add the register info part of IO pad SOC data and
use these register info for accessing PMC register.

This will help to extend the functionalities of IO DPD APIs to
more SoCs.

bug 200259694

Change-Id: I5a43c171de66d5dcb5ae199291976946194dfd84
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1282183
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>