clk: tegra: DFLL regulator undershoot guard-band
Add a mechanism to enforce the regulator undershoot guard-band when
requesting a new DFLL frequency. It is based on the DFLL limits
mechanism we already have.

The value of guard-band is specified by the device tree property
"nvidia,pmic-undershoot-gb'. It can be overriden by writing to the
pmu_undershoot_gb debugfs file.

Based on work by Aleks Frid <afrid@nvidia.com>

Bug 1832505

Change-Id: Iac4fb24a686d458c3685f81a60690e3eabc11bcf
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1282883
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>