soc/tegra: edp: Fix setting for CPU EDP limit via debugfs
When the CPU EDP limit is set via the debugfs, the cpufreq limit is not
updated. Fix this by ensuring the the CPU EDP limit is updated when the
debugfs entry is updated and this change is propagated to the cpufreq
driver. Please note that this only impacts testing when using the
debugfs to force the CPU EDP limit.

Bug 1811732

Change-Id: I05df8b455ea6565c8fed29aed56b77c58e645cb9
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1286843
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>