platform: tegra: kfuse: Remove chip config usage
Remove the usage of chip specific config for compiling the
chip specific code. Instead of use the chip specific data
and decide the feature or control the execution with
compatible data.

This will help in kernel unification to have same defconfig for
all Tegra SoCs.

bug 200257381

Change-Id: I37b9d68877e7c0c0694abb35c3ad99defd335b51
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1291983
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sujeet Baranwal <sbaranwal@nvidia.com>